modified migration that temperature field is a decimal instead of an integer, created model file for weather_information

+++ priv/repo/migrations/20160921142758_add_weather_information.exs
@@ -4,7 +4,7 @@ defmodule GrabbingTheWeather.Repo.Migrations.AddWeatherInformation do
   def change do
     create table(:weather_information) do
       add :city, :string
-      add :temperature, :integer
+      add :temperature, :decimal
       add :description, :string
 
       timestamps
